requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
frozenlist==1.7.0
gitdb==4.0.12
gitpython==3.1.44
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markupsafe==3.0.2
multidict==6.5.0
narwhals==1.42.0
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pip==23.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.1
pyjwt==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
realtime==2.4.3
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.11.3
streamlit==1.45.1
strenum==0.4.15
supabase==2.15.3
supafunc==0.9.4
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websockets==14.2
yarl==1.20.1
